TODO-36 | tests: logout tested
